+++ src/popup/uiHandler.js
@@ -98,6 +98,7 @@ function showAddonMain() {
     document.getElementById("show-only-autograph").style.display = "none";
     document.getElementById("show-only-nokey").style.display = "none";
     document.getElementById("show-only-noint").style.display = "none";
+    document.getElementById("filter-by-tool").style.display = "none";
     document.getElementById("no-tools").style.display = "none";
     document.getElementById("no-indicators").style.display = "none";
     // Hide download icon
@@ -136,6 +137,7 @@ function showButtonsByType(indicator, type, tag, showOnlyFav, showOnlyAutograph,
     document.getElementById("show-only-autograph").style.display = "block";
     document.getElementById("show-only-nokey").style.display = "block";
     document.getElementById("show-only-noint").style.display = "block";
+    document.getElementById("filter-by-tool").style.display = "block";
     document.getElementById("no-tools").style.display = "none";
     document.getElementById("no-indicators").style.display = "none";
     document.getElementById("hist-icon").style.display = "none";